Move over all primitives to a separate crate
