+++ primitives/src/lib.rs
@@ -1,0 +1,84 @@
+// Copyright 2018 Parity Technologies (UK) Ltd.
+// This file is part of Substrate Shasper.
+
+// Substrate is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Substrate is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Substrate.  If not, see <http://www.gnu.org/licenses/>.
+
+#![cfg_attr(not(feature = "std"), no_std)]
+#![cfg_attr(not(feature = "std"), feature(alloc))]
+
+extern crate parity_codec;
+#[macro_use]
+extern crate parity_codec_derive;
+extern crate substrate_primitives as primitives;
+extern crate sr_primitives as runtime_primitives;
+#[macro_use]
+extern crate fixed_hash;
+extern crate sr_std as rstd;
+
+construct_fixed_hash! {
+	/// Fixed 384-bit hash.
+	pub struct H384(48);
+}
+
+pub use primitives::{H256, OpaqueMetadata};
+
+#[cfg(feature = "std")]
+pub use primitives::bytes;
+
+/// Shasper validator public key.
+pub type ValidatorId = primitives::AuthorityId;
+
+/// Alias to Ed25519 pubkey that identifies an account on the chain.
+pub type AccountId = primitives::H256;
+
+/// A hash of some data used by the chain.
+pub type Hash = primitives::H256;
+
+/// Index of a block number in the chain.
+pub type BlockNumber = u64;
+
+/// Index of an account's extrinsic in the chain.
+pub type Nonce = u64;
+
+/// Count value in Shasper.
+pub type Count = u64;
+
+/// Slot value in Shapser.
+pub type Slot = u64;
+
+/// Opaque types. These are used by the CLI to instantiate machinery that don't need to know
+/// the specifics of the runtime. They can then be made to be agnostic over specific formats
+/// of data like extrinsics, allowing for them to continue syncing the network through upgrades
+/// to even the core datastructures.
+pub mod opaque {
+	use super::*;
+	use rstd::prelude::*;
+	use runtime_primitives::{traits::{self, BlakeTwo256}, generic};
+
+	/// Opaque, encoded, unchecked extrinsic.
+	#[derive(PartialEq, Eq, Clone, Default, Encode, Decode)]
+	#[cfg_attr(feature = "std", derive(Serialize, Deserialize, Debug))]
+	pub struct UncheckedExtrinsic(#[cfg_attr(feature = "std", serde(with="bytes"))] pub Vec<u8>);
+	impl traits::Extrinsic for UncheckedExtrinsic {
+		fn is_signed(&self) -> Option<bool> {
+			None
+		}
+	}
+	/// Opaque block header type.
+	pub type Header = generic::Header<BlockNumber, BlakeTwo256, generic::DigestItem<Hash, ValidatorId>>;
+	/// Opaque block type.
+	pub type Block = generic::Block<Header, UncheckedExtrinsic>;
+	/// Opaque block identifier type.
+	pub type BlockId = generic::BlockId<Block>;
+}
