+++ runtime/wasm/Cargo.toml
@@ -11,7 +11,7 @@ integer-sqrt = { git = "https://github.com/paritytech/integer-sqrt-rs.git" }
 safe-mix = { version = "1.0", default-features = false}
 parity-codec-derive = { version = "^2.1" }
 parity-codec = { version = "^2.1", default-features = false }
-substrate-primitives = { git = "https://github.com/paritytech/substrate", default-features = false }
+shasper-primitives = { path = "../../primitives", default-features = false }
 substrate-client = { git = "https://github.com/paritytech/substrate", default-features = false }
 sr-std = { git = "https://github.com/paritytech/substrate", default-features = false }
 sr-io = { git = "https://github.com/paritytech/substrate", default-features = false }
@@ -32,7 +32,7 @@ default = []
 std = [
 	"safe-mix/std",
 	"parity-codec/std",
-	"substrate-primitives/std",
+	"shasper-primitives/std",
 	"substrate-client/std",
 	"sr-std/std",
 	"sr-io/std",
